More work on isolating class functions
