+++ module.json
@@ -10,7 +10,7 @@
       "url": "https://www.buymeacoffee.com/draphtx"
     }
   ],
-  "version": "2.3.6",
+  "version": "2.3.8",
   "minimumCoreVersion" : "10.291",
   "compatibleCoreVersion" : 10,
   "compatibility": {
@@ -18,8 +18,11 @@
       "verified": "10.291"
   },
   "scripts": [
-    "scripts/loadoutsHooks.js",
-    "scripts/config.js"
+    "scripts/init.js",
+    "scripts/config.js",
+    "scripts/loadouts.js",
+    "scripts/hooks.js",
+    "scripts/loadoutDialogs.js"
   ],
   "languages": [
     {
